tests/wpt/tests/html/semantics/forms/the-select-element/stylable-select/select-mouse-behavior.tentative.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://bugs.chromium.org/p/chromium/issues/detail?id=1422275">
<link rel=help href="https://github.com/openui/open-ui/issues/433#issuecomment-1452461404">
<link rel=help href="https://github.com/openui/open-ui/issues/386#issuecomment-1452469497">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>

<style>
select {
  appearance: base-select;
}
</style>

<select id=custombutton>
  <button>button</button>
  <option class=one>one</option>
  <option class=two>two</option>
</select>

<select id=fallbackbutton>
  <option class=one>one</option>
  <option class=two>two</option>
</select>

<script>
for (const id of ['fallbackbutton', 'custombutton']) {
  const select = document.getElementById(id);
  const optionOne = select.querySelector('option.one');
  const optionTwo = select.querySelector('option.two');

  promise_test(async () => {
    assert_false(select.matches(':open'),
      'Select should be closed at the start of the test.');

    await test_driver.click(select);
    assert_true(select.matches(':open'),
      'Select should be open after clicking the button.');

    await test_driver.click(select);
    assert_false(select.matches(':open'),
      'Select should be closed after clicking the button a second time.');
  }, `${id}: Select with appearance:base-select should open and close when clicking the button.`);

  promise_test(async () => {
    assert_false(select.matches(':open'),
      'Select should be closed at the start of the test.');
    assert_equals(select.value, 'one',
      'Select.value should be one at the start of the test.');

    await test_driver.click(select);
    assert_true(select.matches(':open'),
      'Select should be open after clicking the button.');

    await test_driver.click(optionTwo);
    assert_false(select.matches(':open'),
      'Select should be closed after clicking an option.');
    assert_equals(select.value, 'two',
      'Select.value should be two after clicking the option.');
  }, `${id}: Clicking an option in an appearance:base-select select should choose the option and close the popover.`);
}
</script>
